requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
pillow==10.1.0
dj-database-url==2.1.0
whitenoise==6.6.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
tqdm==4.66.1 